requirements.txt
python-decouple
jira